その時点のメモの状態とメモの各種操作のメソッドを取りまとめるコントローラークラスを導入
